Unbreak the build on 64-bit architectures as well.

+++ sys/dev/usb/if_rum.c
@@ -1192,7 +1192,7 @@ rum_tx_mgt(struct rum_softc *sc, struct mbuf *m0, struct ieee80211_node *ni)
 		xferlen += 4;
 
 	DPRINTFN(10, ("sending mgt frame len=%d rate=%d xfer len=%d\n",
-	    m0->m_pkthdr.len + RT2573_TX_DESC_SIZE, rate, xferlen));
+	    m0->m_pkthdr.len + (int)RT2573_TX_DESC_SIZE, rate, xferlen));
 	
 	usbd_setup_xfer(data->xfer, sc->sc_tx_pipeh, data, data->buf, xferlen,
 	    USBD_FORCE_SHORT_XFER | USBD_NO_COPY, RUM_TX_TIMEOUT, rum_txeof);
@@ -1354,7 +1354,7 @@ rum_tx_data(struct rum_softc *sc, struct mbuf *m0, struct ieee80211_node *ni)
 		xferlen += 4;
 
 	DPRINTFN(10, ("sending frame len=%d rate=%d xfer len=%d\n",
-	    m0->m_pkthdr.len + RT2573_TX_DESC_SIZE, rate, xferlen));
+	    m0->m_pkthdr.len + (int)RT2573_TX_DESC_SIZE, rate, xferlen));
 
 	usbd_setup_xfer(data->xfer, sc->sc_tx_pipeh, data, data->buf, xferlen,
 	    USBD_FORCE_SHORT_XFER | USBD_NO_COPY, RUM_TX_TIMEOUT, rum_txeof);
